# emissor_nfe/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Emissor NF-e API
  version: 1.0.0
  description: |
    API REST multi-empresa para emissão de NF-e (modelo 55) com certificado A1,
    autorização, cancelamento, carta de correção, inutilização, consulta e DANFE.
servers:
  - url: http://localhost:8000/api/v1
    description: Local
...
paths:
...
  /empresas:
    get:
      tags: [Empresas]
      security: [{ bearerAuth: [] }]
      summary: Listar empresas do usuário
      responses:
        '200': { description: Lista }
    post:
      tags: [Empresas]
      security: [{ bearerAuth: [] }]
      summary: Cadastrar empresa (CNPJ)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpresaInput'
      responses:
        '201': { description: Criada }
...
  /empresas/{empresa}/nfe/{chave}/xml:
    get:
      tags: [NFe]
      security: [{ bearerAuth: [] }]
      summary: Download XML
      parameters:
        - $ref: '#/components/parameters/empresaId'
        - $ref: '#/components/parameters/chave'
      responses:
        '200':
          description: XML
          content:
            application/xml:
              schema: { type: string }
  /empresas/{empresa}/nfe/{chave}/danfe:
    get:
      tags: [NFe]
      security: [{ bearerAuth: [] }]
      summary: Download DANFE PDF
      parameters:
        - $ref: '#/components/parameters/empresaId'
        - $ref: '#/components/parameters/chave'
      responses:
        '200':
          description: PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  parameters:
    empresaId:
      name: empresa
      in: path
      required: true
      schema: { type: integer }
    chave:
      name: chave
      in: path
      required: true
      schema: { type: string, minLength: 44, maxLength: 44 }
  schemas:
    EmpresaInput:
      type: object
      required:
        - cnpj
        - razao_social
        - logradouro
        - numero
        - bairro
        - municipio
        - codigo_municipio
        - uf
        - cep
        - crt
      properties:
        cnpj: { type: string, minLength: 14, maxLength: 14 }
        ie: { type: string }
        razao_social: { type: string }
        nome_fantasia: { type: string }
        email: { type: string }
        telefone: { type: string }
        logradouro: { type: string }
        numero: { type: string }
        complemento: { type: string }
        bairro: { type: string }
        municipio: { type: string }
        codigo_municipio: { type: string, minLength: 7, maxLength: 7 }
        uf: { type: string, minLength: 2, maxLength: 2 }
        cep: { type: string, minLength: 8, maxLength: 8 }
        crt: { type: integer, enum: [1, 2, 3] }
        ambiente: { type: string, enum: [homologacao, producao] }
        serie_inicial: { type: integer }
        proximo_numero: { type: integer }
...